hello world express server setup
Setting up a "Hello, World!" express server with docker &
docker-compose. Had some issue with starting the server and the
container in which it was deployed. Removed and re-created the container
image, and it started working fine. No idea what the issue was, but I'm
guessing it was to do with either/or:
* Dockerfile build commands while creating the original image
* The volumes linked to the container
* The original image
* Some arbitraty limitation of docker, that I have yet to learn